requirements.txt
asgiref==3.2.5
dj-static==0.0.6
Django==3.0.4
djangorestframework==3.11.0
mysqlclient==1.4.6
passlib==1.7.2
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.1
static3==0.7.0
